latest updated version of minishell that was pushed in master

+++ parsing/builtins/check_word.c
@@ -6,11 +6,11 @@
 /*   By: rennacir <rennacir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/14 08:25:17 by rennacir          #+#    #+#             */
-/*   Updated: 2023/07/26 15:57:11 by rennacir         ###   ########.fr       */
+/*   Updated: 2023/08/08 22:58:57 by rennacir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../minishell.h"
+#include "../../minishell.h"
 
 int	check_word(char *mai, char *str)
 {
@@ -19,7 +19,7 @@ int	check_word(char *mai, char *str)
 	i = 0;
 	while (mai[i] && str[i] && (mai[i] == str[i] || mai[i] == str[i] + 32))
 		i++;
-	if (i == ft_strlen(mai))
+	if (i == (int)ft_strlen(mai))
 		return (1);
 	return (0);
 }
